requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmdstanpy==1.0.8
contourpy==1.0.6
convertdate==2.4.0
cryptography==38.0.3
cycler==0.11.0
Cython==0.29.32
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
ephem==4.1.3
fastapi==0.85.2
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.1
h11==0.14.0
hijri-converter==2.2.4
holidays==0.16
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
matplotlib==3.6.2
motor==3.1.1
multidict==6.0.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
passlib==1.7.4
Pillow==9.3.0
pip==22.2.2
pipenv==2022.10.25
platformdirs==2.5.2
prophet==1.1.1
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyMeeus==0.5.11
pymongo==4.3.2
pyparsing==3.0.9
pystan==2.19.1.1
python-dateutil==2.8.2
python-dotenv == 0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
requests==2.28.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
setuptools==65.5.0
setuptools-git==1.2
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
virtualenv==20.16.6
virtualenv-clone==0.5.7
wheel==0.37.1
yarl==1.8.1
